requirements.txt
msal==1.24.1
requests==2.31.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-search-documents==11.4.0
tenacity==8.2.3
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
click==8.1.7